# lib/api-zod/src/generated/types/contentBlueprint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentBlueprintAlgorithmReadinessBreakdown } from "./contentBlueprintAlgorithmReadinessBreakdown";

export interface ContentBlueprint {
  analysisId: number;
  algorithmReadinessScore: number;
  algorithmReadinessBreakdown: ContentBlueprintAlgorithmReadinessBreakdown;
  characterConsistencyTips: string[];
  soundDesignPlan: string[];
  povLoreIdea: string;
  identityLoyaltyFactors: string[];
  highIntentGainsTactics: string[];
}